fix(guards): comment-aware scanning per the #386 ruling — audit + shared tokeniser
The British-English copy guard scanned raw source INCLUDING comments/JSDoc, so a
design-note that merely mentioned an American spelling reddened CI. ceeRecovery.ts
had to misspell its own example `analyze_sentiment` (commit 49e8c53e) to dodge it —
the identical footgun PR #386 fixed in the alpha-emission guard by stripping
comments before scanning.

- Extract #386's literal-aware comment stripper verbatim into a shared module
  (tests/helpers/stripSourceComments.ts) and route BOTH guards through it, so the
  two cannot drift apart (derive-don't-mirror; the inline copy is removed from
  semantic-colour-alpha.spec.ts).
- british-english.spec.ts strips comments (literal-aware; strings/JSX survive)
  before the scan. Adds a both-directions mutation-proof describe block.
- Restore ceeRecovery.ts's JSDoc to the natural bare `analyze` as living proof.

Mutation-checked: with the strip disabled the positive control (an American
spelling in a string literal) still fails RED — the guard is not vacuous — while
the three comment-only cases correctly go RED, proving they depend on the strip.
